Convert to polkit from usermode consolehelper for gaining root privileges
polkit is better integrated into various environments, both console and GUI
and offers better access rules and prevents the internal need to run
su which is prone to errors and doesn't offer an environment agnostic
prompt to the user.

In this case the current package policy (in rpm spec) has been migrated
here and is as follows:

drakclock requires no authentication to run (just a console login)

drakfont requires authentication as the current user.

drakups, drakauth, draklog and drakxservices require authentication as and
administrator.

Conflicts:
	perl-install/Makefile.drakxtools