reverse the if branches
rationale:
since we no more offer an OSS alternative, the list is empty and thus
we were not offering the dialog anymore.

However it does more than offering to switch between OSS & ALSA: we do
want to offer to configure PA even if we cannot switch to another
driver.

Thus the GUI is now either show an error dialog if there's no
supported sound card of showing PA options