fix selecting of uneeded locales (mga#3723)
original fix by Pablo Saratxaga

Mageia has he ability to choose both *language* and *country*
Language is used for choosing translations to be displayed, etc.
country for things like currency, etc.
(country may be deduced from other geographical settings rather than
choosen directly, but that is not the problem).

GNU libc locales embedd both concepts (language and country) in a single
locale.
Obviously those exists for only some combinations.

For cases were an existing matching locale doesn't exists, Mageia
chooses two locales, one for the language oriented settings, and another
four the country oriented settings.

Unlike normal installer where packages to be installed are choosen; in
draklive-install, we have a set of "all packages", and "uneeded" ones
are removed.
The problem was obviously the installer didn't see the "need" for the
locales-xx package for country; only language was checked.

This fixes that, the idea is to check for *both* $locale->{lang} and
$locale->{country}, which can lead two *two* neeeded locales-xx packages